tried to add text movement animation in animations.dart
